Postgre module v26,SimpleQuery task result subscriber and message decode

+++ jdbd-postgre/src/main/java/io/jdbd/postgre/protocol/client/Messages.java
@@ -15,7 +15,7 @@ abstract class Messages {
 
     static final byte STRING_TERMINATOR = '\0';
 
-    static final int LENGTH_SIZE = 4;
+    static final byte LENGTH_SIZE = 4;
 
     /**
      * <ul>
